Pushed last changes

+++ crypto/ec_sign.c
@@ -25,7 +25,8 @@ uint8_t *ec_sign(EC_KEY const *key, uint8_t const *msg, size_t msglen,
 	if (!key || !msg || !sig)
 		return (NULL);
 
-	if (ECDSA_sign(0, msg, msglen, sig->sig, &(sig->len), (EC_KEY *)key) != 1)
+	if (ECDSA_sign(0, msg, msglen, sig->sig, &(unsigned int)(sig->len),
+				(EC_KEY *)key) != 1)
 		return (NULL);
 
 	return (sig->sig);
